requirement.txt
langchain
langchain_community
langchain-core
python-dotenv
streamlit
pypdf
langchain-text-splitters
sentence_transformers
langchain_huggingface
chromadb
langchain_groq
bs4
arxiv
pymupdf
faiss-cpu
langchain_chroma





